Fix queen,knight,king & pawn bugs and make black pieces functional
Our chess game is far from complete. I tried to the solve the following problems today but failed to make it work.... when a piece is clicked,no other pieces cant be clicked, set the valid moves for possible capture, player turn change and correcting the legal moves relative to the pieces enemy and friend positions on the board.